# docs/swagger.yaml
definitions:
  models.ErrorMessage:
    properties:
      error:
        type: string
    type: object
...
  models.Language:
    properties:
      created:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      name:
        type: string
      updated:
        type: string
    required:
    - created_by
    - name
    type: object
  models.Project:
    properties:
      created_by:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      language:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    required:
    - created_by
    - language
    - name
    type: object
...
  models.SuccessFindLanguage:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
  models.SuccessFindLanguages:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Project'
        type: array
    type: object
  models.SuccessFindProject:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
  models.SuccessFindProjects:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Project'
        type: array
    type: object
...
  models.SuccessLanguageCreation:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
  models.SuccessLanguageDelete:
    properties:
      data:
        type: boolean
    type: object
  models.SuccessLanguageUpdate:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
  models.SuccessProjectCreation:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
  models.SuccessProjectDelete:
    properties:
      data:
        type: boolean
    type: object
  models.SuccessProjectUpdate:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
...
info:
  contact: {}
paths:
...
  /languages:
    get:
      description: Bulk all nepackage's languages
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessFindLanguages'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: FindLanguages
      tags:
      - Languages
    post:
      description: Create language with models.Language model
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Language data
        in: body
        name: language
        required: true
        schema:
          $ref: '#/definitions/models.Language'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessLanguageCreation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: CreateLanguage
      tags:
      - Languages
  /languages/{id}:
    delete:
      description: Delete language with using id
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Language ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessLanguageDelete'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: DeleteLanguage
      tags:
      - Languages
    get:
      description: Find a language with given id in path
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Language ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessFindLanguage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: FindLanguage
      tags:
      - Languages
    patch:
      description: Update language with models.Language model
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Language data
        in: body
        name: language
        required: true
        schema:
          $ref: '#/definitions/models.Language'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessLanguageUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: UpdateLanguage
      tags:
      - Languages
  /projects:
    get:
      description: Bulk all nepackage's projects
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessFindProjects'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: FindProjects
      tags:
      - Projects
    patch:
      description: Update project with models.Project model
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessProjectUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: UpdateProject
      tags:
      - Projects
    post:
      description: Create project with models.Project model
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessProjectCreation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: CreateProject
      tags:
      - Projects
  /projects/{id}:
    delete:
      description: Delete project using id
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessProjectDelete'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: DeleteProject
      tags:
      - Projects
    get:
      description: Find a project with given id in path
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessFindProject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: FindProject
      tags:
      - Projects
...
swagger: "2.0"